% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cohort.R
\name{compute_values}
\alias{compute_values}
\title{Compute State Values per Cycle}
\usage{
compute_values(x)
}
\arguments{
\item{x}{An object of class \code{eval_model}.}
}
\value{
A data.frame of state values, one column per
  state value and one row per cycle.
}
\description{
Given an evaluated model, computes the total state values
per cycle.
}

