% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{define_matrix}
\alias{define_matrix}
\title{Title}
\usage{
define_matrix(..., state_names = LETTERS[seq_len(sqrt(length(list(...))))])
}
\arguments{
\item{...}{}

\item{state_names}{}
}
\description{
Title
}

