% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{define_model}
\alias{define_model}
\title{Define a Markov Model}
\usage{
define_model(parameters = define_parameters(),
  transition_matrix = define_matrix(), states = define_state_list())
}
\arguments{
\item{parameters}{Optional. An object generated by 
\link{\code{define_parameters}}.}

\item{transition_matrix}{An object generated by 
\link{\code{define_matrix}}.}

\item{states}{An object generated by 
\link{\code{define_states}}.}
}
\value{
An object of class \code{uneval_model} (a list 
  containing the unevaluated parameters, matrix and
  states).
}
\description{
Combine information on parameters, transition matrix and 
states defined through \link{\code{define_parameters}}, 
\link{\code{define_matrix}} and 
\link{\code{define_states}} respectively.
}
\details{
This function checks whether the objects are compatible
in the same model (same state names...).
}
\examples{

param <- define_parameters(
  a = 1,
  b = 4 * markov_cycle,
  c = a + b
)

mat <- define_matrix(
  state_names = c("s1", "s2")
  1 / c, 1 - 1/ c,
  0, 1
)

list_states <- define_list_states(
  s1 = state(cost = 234),
  s2 = state(cost = 421)
)

define_model(
 parameters = param,
 transition_matrix = mat,
 states = list_states
)

}

