% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{eval_states}
\alias{eval_states}
\title{Evaluate Markov Model States}
\usage{
eval_states(x, parameters)
}
\arguments{
\item{x}{An \code{uneval_states} object generated by 
\code{\link{define_state_list}}.}

\item{parameters}{An \code{eval_parameters} object 
generated by \code{\link{eval_parameters}}.}
}
\value{
An \code{eval_states} object, a list with one
  data.frame per state containing a column per state
  value and a line per cycle.
}
\description{
Evaluate Markov Model States
}

