% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_counts}
\alias{get_counts}
\title{Get Markov Model Counts}
\usage{
get_counts(x)
}
\arguments{
\item{x}{An \code{eval_model} object.}
}
\value{
A \code{cycle_counts} object.
}
\description{
Only works on evaluated models.
}
\details{
For internal use.
}

