% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_parameters}
\alias{get_parameters}
\title{Get Markov Model Parameters}
\usage{
get_parameters(x)
}
\arguments{
\item{x}{An \code{uneval_model} or \code{eval_model}
object.}
}
\value{
An \code{uneval_parameters} or
  \code{eval_parameters} object.
}
\description{
Works on both unevaluated and evaluated models.
}
\details{
For internal use.
}

