% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_states}
\alias{get_states}
\title{Get Markov Model States}
\usage{
get_states(x)
}
\arguments{
\item{x}{An \code{uneval_model} or \code{eval_model}
object.}
}
\value{
An \code{uneval_states} or \code{eval_states}
  object.
}
\description{
Works on both unevaluated and evaluated models.
}
\details{
For internal use.
}

