% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{modify}
\alias{modify}
\title{Modify Object}
\usage{
modify(x, ...)
}
\arguments{
\item{x}{Various objects.}

\item{...}{Modifications.}
}
\value{
Same class as \code{x}.
}
\description{
This generic function allows the modification of various 
objects such as parameters, transitions matrix or states.
}
\details{
More details are available on the respective help page of
each object definition.
}

