% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{names_parameters}
\alias{names_parameters}
\title{Return parameters names}
\usage{
names_parameters(x)
}
\arguments{
\item{x}{\code{uneval_parameters} or
\code{eval_parameters} object.}
}
\value{
A character vector of parameter names.
}
\description{
Extract parameters names from an \code{uneval_parameters}
or \code{eval_parameters} object.
}

