% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_models}
\alias{run_model}
\alias{run_models}
\title{Run one or more Markov Model}
\usage{
run_models(..., init, cycles, count_args = NULL)

run_model(...)
}
\arguments{
\item{...}{\code{uneval_model} objects.}

\item{init}{numeric vector, same length as number of 
model states. Number of individuals in each model state
at the beginning.}

\item{cycles}{positive integer. Number of Markov Cycles 
to compute.}

\item{count_args}{Additional arguments passed as a list 
to \code{compute_counts}.}
}
\value{
A list of individual counts and state values per 
  cycle and total state values.
}
\description{
Runs a set of unevaluated Markov Models. All models 
should have the same states and state value names.
}
\details{
A classical situation where more than one model needs to 
be run is to compare different care startegies.

In order to compute comparisons Markov Models must be 
similar (same states and state value names). Thus models 
should only differ through parameters, transition matrix 
cell values and values attached to states (but not value 
names).

The initial number of individuals in each model state and
the number of cycle will be the same for all models.

This function can run only one model.
}

