% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cohort.R
\name{simulate_cohort}
\alias{simulate_cohort}
\title{Simulate Markov Model}
\usage{
simulate_cohort(model, init, cycles, ...)
}
\arguments{
\item{model}{An \code{uneval_model} object.}

\item{init}{numeric vector, same length as number of 
model states. Number of individuals in each model state
at the beginning.}

\item{cycles}{positive integer. Number of Markov Cycles 
to compute.}

\item{...}{Additional arguments passed to
\code{compute_counts}.}
}
\value{
A list of individual counts and state values per
  cycle and total state values.
}
\description{
This function name should change.
}
\details{
Examples should be given.
}

