% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilist.R
\name{simulate_cohort_iter}
\alias{simulate_cohort_iter}
\title{Simulate Iteratively a Markov Model Over New Parameter
Values}
\usage{
simulate_cohort_iter(model, cycles, init, newdata)
}
\arguments{
\item{model}{An \code{uneval_model} object.}

\item{cycles}{positive integer. Number of Markov Cycles 
to compute.}

\item{init}{numeric vector, same length as number of 
model states. Number of individuals in each model state
at the beginning.}

\item{newdata}{a data.frame whose names match parameters
names. \code{model} will be evaluated iteratively,
taking successivel values from each row.}
}
\value{
A data.frame containing the values of \code{newdata} and each Markov Model evaluation in \code{res}.
}
\description{
This function name should change.
}

