% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{state_count}
\alias{state_count}
\title{Return Number of State}
\usage{
state_count(x)
}
\arguments{
\item{x}{An object containing states.}
}
\value{
An integer: number of states.
}
\description{
For internal use.
}
\details{
Work with both \code{uneval_states} and \code{eval_states}.
}

